Multimodal RAG Pipeline/requirements.txt
# ============================================
# Multimodal RAG — Dependencies
# ============================================
# Install:  pip install -r requirements.txt
# ============================================

# --- Core Framework ---
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# --- Vector Store ---
qdrant-client>=1.12.0
langchain-qdrant>=0.2.0

# --- PDF Processing ---
pymupdf>=1.25.0            # Fast PDF text + image + table extraction
pandas>=2.0.0              # Table extraction to dataframes

# --- Embeddings & LLM ---
openai>=1.60.0

# --- Web UI ---
streamlit>=1.40.0

# --- Utilities ---
python-dotenv>=1.0.0
Pillow>=10.0.0
